Support tables with foreign key constraints #161
Ghostferry currently documents that it cannot work with foreign-key
constraints (FKCs). This is because data- and binlog-writers operate
at the same time and copying happens on tables in parallel, which may
insert data that (temporarily) conflicts constraints.

MySQL allows disabling foreign-key constraint checks on a per-session
basis, and it does not re-validate constraints when this is disabled.
As a result, we may may temporarily disable constraint enforcement until
the database is back in a consistent state. The only issue that does
arise is that tables must be created in an order that satisfies their
inter-dependencies.

To disable the FKCs, one must add the following config to the target DB
configuration:

  "Params": {
    "foreign_key_checks": "0"
  }

Change-Id: I61ffcca56d6bc5517e961bfd5f5df4e02878b557